# docs/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rust-web-dashboard",
    "description": "Standalone Rust web dashboard for a DCS-gRPC server (single-binary axum + tonic)",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/atmosphere": {
      "get": {
        "tags": [
          "dcs"
        ],
        "summary": "`GET /api/atmosphere?lat=&lon=&alt=` → wind + temperature/pressure\n(AtmosphereService). `alt` defaults to 0, matching the source.",
        "operationId": "atmosphere",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "format": "double"
            }
          },
          {
            "name": "alt",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Atmosphere data"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
    "/api/console": {
      "post": {
        "tags": [
          "dcs"
        ],
        "summary": "`POST /api/console` → evaluate Lua in the mission (CustomService.Eval).",
        "operationId": "console",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsoleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Lua evaluation result"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
    "/api/srs/clients": {
      "get": {
        "tags": [
          "srs"
        ],
        "operationId": "get_clients",
        "responses": {
          "200": {
            "description": "List of SRS clients"
          },
          "503": {
            "description": "Connected-client data temporarily unavailable"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/srs/settings": {
      "get": {
        "tags": [
          "srs"
        ],
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "SRS configuration settings"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      },
      "post": {
        "tags": [
          "srs"
        ],
        "operationId": "post_settings",
        "requestBody": {
          "description": "New SRS configuration settings",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SRS configuration updated"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsoleBody": {
        "type": "object",
        "properties": {
          "lua": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}